Moved webserver into class so can test
- Also renamed to server instead of program
